Allow lists of objects in parser
Support for lists of maps/objects was added to the golang HCL parser in
June 2016, and the test case list_of_maps.hcl was added for it. Add
support to the PyHCL parser, lift this test case, and also enable test
case assign_deep.hcl, which can now be expected to pass.